Let the GSource dispatch python callbacks
Turn things around and rather than iterating the GLib main context from
python iterate the python mainloop from GLib.

In order to do this, we need to be able to calculate the timeout (and
whether anything can be dispatched) from inside the prepare and check
functions of the source.
We can do that easily by looking at the _ready and _schedule attributes
of the loop. Once we have that, we can dispatch everything scheduled by
python by calling _run_once and relying on our select() implementation
to return immediately.